Cannot address both string or number
Add `string` and `number` types support for built-in Base Components.

Instead of passing the event target value to the `onChange` functions we extract the values from the options passed to the component. 

This way we do not have to care about the type of the value. Internally the value of each option is cast into a `String`  for comparison, so values must still be stringable.