Introduce Enriched Control Option
The changes to the value object will let it carry on extra data among
the business logic.

This permit third parties to fill extra information to consume during
presentation.

Due to the need to include an additional Data Structure the changes are
involving the re-organization of the Value Objects and the Data
Structures, previously mixed under the `vo` directory.